requirements.txt
django
djangorestframework
djangorestframework-simplejwt
requests
python-dotenv
pytest
factory-boy